Seperate Stage Widget to the new class
